Added all files to repo

+++ adventure_works/models/staging/Sales/stg_salesorderdetail.sql
@@ -12,8 +12,8 @@ with
             , productid as fk_produto
             , specialofferid as fk_oferta_especial
             , carriertrackingnumber as numero_rastreamento_transportadora
-            , cast(orderqty as quantidade
-            , cast(unitprice as numeric(10,2)) as preco_unitario
+            , cast(orderqty as int) as quantidade
+            , cast(unitprice as numeric(15,4)) as preco_unitario
             , cast(unitpricediscount as numeric(5,2)) as desconto
             , cast(modifieddate as date) as data_alteracao
             , rowguid
